# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NBE10-12-2-Team1 API
  description: |
    오늘 뭐먹지 SNS 서비스 REST API 문서입니다.

    ## 인증
    인증이 필요한 API는 요청 헤더에 Bearer 토큰을 포함해야 합니다.
    ```
    Authorization: Bearer {accessToken}
    ```
  version: 1.0.0
servers:
  - url: https://api.example.com
    description: Production Server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/restaurants/{id}:
    get:
      tags: [Restaurants]
      summary: 식당 상세 조회
      description: 식당 ID로 식당 상세 정보를 조회합니다.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: 조회할 식당 ID
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: 식당 상세 조회 성공
          content:
            application/json:
              example:
                success: true
                data:
                  id: 1
                  kakaoPlaceId: "12345678"
                  name: "을지삼겹"
                  category: "한식"
                  address: "서울 중구 을지로3가 12-3"
                  roadAddress: "서울 중구 을지로 123길 4"
                  region1: "서울"
                  region2: "중구"
                  region3: "을지로3가"
                  phone: "02-123-4567"
                  lat: 37.5665
                  lng: 126.9780
                  createdAt: "2026-06-18T17:00:00"
        "404":
          description: 존재하지 않는 식당 (RESTAURANT_NOT_FOUND)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "RESTAURANT_NOT_FOUND"
                  message: "존재하지 않는 식당입니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."

  # ─────────────────────────────────────────────
  # Feeds
  # ─────────────────────────────────────────────
...
  /api/v1/lists:
    post:
      tags: [Lists]
      summary: 맛집 리스트 생성
      description: 새로운 맛집 리스트를 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  maxLength: 200
                  example: "을지로 감성 데이트 코스"
                description:
                  type: string
                  maxLength: 500
                  example: "좁은 골목길 숨겨진 보석 같은 감성 맛집 모음집"
                moodTag:
                  type: string
                  enum: ["데이트", "혼밥", "회식", "야식"]
                  example: "데이트"
      responses:
        "201":
          description: 맛집 리스트 생성 성공
          content:
            application/json:
              example:
                success: true
                data:
                  listId: 10
                  userId: 1
                  title: "을지로 감성 데이트 코스"
                  description: "좁은 골목길 숨겨진 보석 같은 감성 맛집 모음집"
                  moodTag: "데이트"
                  createdAt: "2026-06-18T17:52:00"
                message: "맛집 리스트가 성공적으로 생성되었습니다."
        "400":
          description: title 비어있음 / 길이 초과 / moodTag 허용값 아님 (INVALID_MOOD_TAG)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_MOOD_TAG"
                  message: "허용되지 않는 분위기 태그입니다. (데이트 / 혼밥 / 회식 / 야식)"
        "401":
          description: 로그인하지 않은 사용자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "로그인 후 이용해주세요."
        "500":
          description: 리스트 저장 실패
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."

    get:
      tags: [Lists]
      summary: 맛집 리스트 검색/목록 조회
      description: 맛집 리스트를 검색하거나 목록을 조회합니다.
      parameters:
        - name: q
          in: query
          required: false
          description: 검색어 (제목 + 설명)
          schema:
            type: string
            example: "을지로"
        - name: mood
          in: query
          required: false
          description: 분위기 태그 필터
          schema:
            type: string
            enum: ["데이트", "혼밥", "회식", "야식"]
            example: "데이트"
        - name: userId
          in: query
          required: false
          description: 특정 유저 리스트만 조회
          schema:
            type: integer
            example: 1
        - name: page
          in: query
          required: false
          description: 페이지 번호 (기본값 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          description: 페이지 크기 (기본값 10)
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: 맛집 리스트 목록 조회 성공
          content:
            application/json:
              example:
                success: true
                data:
                  lists:
                    - listId: 10
                      userId: 1
                      nickname: "푸디보이"
                      title: "을지로 감성 데이트 코스"
                      moodTag: "데이트"
                      itemCount: 3
                      createdAt: "2026-06-18T17:52:00"
                  totalPages: 5
                  totalElements: 42
        "400":
          description: userId 형식 오류 또는 page/size가 0 이하 (INVALID_INPUT_VALUE / INVALID_PAGE_SIZE)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_INPUT_VALUE"
                  message: "입력값이 올바르지 않습니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
  /api/v1/lists/{id}:
    get:
      tags: [Lists]
      summary: 맛집 리스트 상세 조회 (아이템 포함)
      description: 특정 맛집 리스트의 상세 정보와 아이템 목록을 조회합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 리스트 ID
          schema:
            type: integer
            example: 10
      responses:
        "200":
          description: 리스트 상세 조회 성공
          content:
            application/json:
              example:
                success: true
                data:
                  listId: 10
                  userId: 1
                  creatorNickname: "푸디보이"
                  title: "을지로 감성 데이트 코스"
                  description: "좁은 골목길 숨겨진 보석 같은 감성 맛집 모음집"
                  moodTag: "데이트"
                  items:
                    - itemId: 30
                      restaurantId: 1
                      restaurantName: "을지삼겹"
                      category: "한식"
                      orderIndex: 1
                      memo: "삼겹살에 미나리 조합은 이길 수가 없음"
                      createdAt: "2026-06-18T17:53:00"
                  createdAt: "2026-06-18T17:52:00"
        "400":
          description: listId 형식 오류 (INVALID_INPUT_VALUE)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_INPUT_VALUE"
                  message: "입력값이 올바르지 않습니다."
        "404":
          description: 존재하지 않는 리스트 (LIST_NOT_FOUND)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "LIST_NOT_FOUND"
                  message: "존재하지 않는 맛집 리스트입니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
  /api/v1/lists/{id}/items:
    post:
      tags: [Lists]
      summary: 리스트에 식당 추가
      description: 맛집 리스트에 식당을 추가합니다. 본인만 가능합니다. 리스트당 최대 5,000개까지 등록 가능합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 리스트 ID
          schema:
            type: integer
            example: 10
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [restaurantId]
              properties:
                restaurantId:
                  type: integer
                  example: 1
                memo:
                  type: string
                  maxLength: 200
                  example: "삼겹살에 미나리 조합은 이길 수가 없음"
      responses:
        "201":
          description: 식당 추가 성공
          content:
            application/json:
              example:
                success: true
                data:
                  itemId: 30
                  listId: 10
                  restaurantId: 1
                  restaurantName: "을지삼겹"
                  memo: "삼겹살에 미나리 조합은 이길 수가 없음"
                message: "리스트에 식당이 추가되었습니다."
        "400":
          description: restaurantId 없음 / memo 200자 초과 / 아이템 5000개 초과 (LIST_ITEM_LIMIT_EXCEEDED)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "LIST_ITEM_LIMIT_EXCEEDED"
                  message: "리스트에 추가할 수 있는 식당은 최대 5,000개입니다."
        "401":
          description: 로그인하지 않은 사용자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "로그인 후 이용해주세요."
        "403":
          description: 리스트 작성자가 아님
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "FORBIDDEN"
                  message: "리스트 작성자만 식당을 추가할 수 있습니다."
        "404":
          description: 리스트 또는 식당이 존재하지 않음 (LIST_NOT_FOUND / RESTAURANT_NOT_FOUND)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "LIST_NOT_FOUND"
                  message: "존재하지 않는 맛집 리스트입니다."
        "409":
          description: 이미 해당 리스트에 같은 식당 존재 (DUPLICATE_LIST_ITEM)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "DUPLICATE_LIST_ITEM"
                  message: "이미 리스트에 추가된 식당입니다."
        "500":
          description: 리스트 아이템 저장 실패
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
  /api/v1/lists/{id}/items/{itemId}:
    put:
      tags: [Lists]
      summary: 리스트 내 식당 정보 수정 (순서/메모)
      description: 리스트 내 식당 아이템의 순서나 메모를 수정합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 리스트 ID
          schema:
            type: integer
            example: 10
        - name: itemId
          in: path
          required: true
          description: 아이템 ID
          schema:
            type: integer
            example: 30
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderIndex:
                  type: integer
                  example: 2
                memo:
                  type: string
                  maxLength: 200
                  example: "여기 볶음밥도 필수 코스입니다."
      responses:
        "200":
          description: 아이템 수정 성공
          content:
            application/json:
              example:
                success: true
                data:
                  itemId: 30
                  orderIndex: 2
                  memo: "여기 볶음밥도 필수 코스입니다."
                message: "리스트 아이템 정보가 변경되었습니다."
        "400":
          description: memo 200자 초과
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_INPUT_VALUE"
                  message: "메모는 200자를 초과할 수 없습니다."
        "401":
          description: 로그인하지 않은 사용자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "로그인 후 이용해주세요."
        "403":
          description: 리스트 작성자가 아님
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "FORBIDDEN"
                  message: "리스트 작성자만 수정할 수 있습니다."
        "404":
          description: 존재하지 않는 아이템 (LIST_ITEM_NOT_FOUND)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "LIST_ITEM_NOT_FOUND"
                  message: "존재하지 않는 리스트 아이템입니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."

    delete:
      tags: [Lists]
      summary: 리스트에서 식당 제외
      description: 리스트에서 식당 아이템을 제외합니다. 본인만 가능합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 리스트 ID
          schema:
            type: integer
            example: 10
        - name: itemId
          in: path
          required: true
          description: 아이템 ID
          schema:
            type: integer
            example: 30
      responses:
        "200":
          description: 식당 제외 성공
          content:
            application/json:
              example:
                success: true
                data: null
                message: "리스트에서 식당이 제외되었습니다."
        "401":
          description: 로그인하지 않은 사용자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "로그인 후 이용해주세요."
        "403":
          description: 리스트 작성자가 아님
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "FORBIDDEN"
                  message: "리스트 작성자만 삭제할 수 있습니다."
        "404":
          description: 존재하지 않는 아이템 (LIST_ITEM_NOT_FOUND)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "LIST_ITEM_NOT_FOUND"
                  message: "존재하지 않는 리스트 아이템입니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
...